src/a2a_servers/pyproject.toml
[project]
name = "aifoundry-a2a-demo"
version = "0.1.0"
description = "Config-driven multi-agent A2A server backed by Azure AI Foundry"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.0",
    "aiohttp>=3.9.0",
    "azure-ai-projects>=2.0.0b4",
    "azure-identity>=1.23.0",
    "azure-storage-blob>=12.0.0",
    "openai>=1.79.0",
    "uvicorn>=0.34.2",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "starlette>=0.35.0",
    "httpx>=0.25.0",
    "sse-starlette>=2.3.5",
]

[dependency-groups]
dev = [
  "ruff==0.15.0",
  "deptry>=0.24.0",
  "pre-commit>=4.5.1",
  "pytest>=9.0.2",
  "pytest-asyncio>=1.3.0",
  "mypy>=1.19.1",
]

[tool.deptry]
known_first_party = [
  "a2a_servers",
  "agent_definition",
  "app_factory",
  "foundry_agent",
  "foundry_agent_executor",
  "settings",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".."]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
  "--strict-markers",
  "--strict-config",
]
